# internal/http/api/openapi.yaml
openapi: 3.0.3
info:
  title: MeshSDR hub API
  version: "1"
  description: |
    JSON API of the MeshSDR hub (TECHNICAL_SPEC §6.10, ADR 0023): only
    what scripts, islands, nodes and operations need. UI actions are HTML
    forms on page-scoped paths, with no API twin. Every endpoint is
    declared here first, then the Go server interface is generated with
    oapi-codegen. Errors use one format: RFC 9457 problem details with a
    stable `code`.

    Every operation declares its required role in `x-meshsdr-access`
    (`anonymous`, `listener`, `operator` or `admin`): the server-side policy
    table (SR-15). A role implies the roles below it; `admin` operations are
    also restricted to `admin.allowed_networks`.

    State-changing requests (POST, PUT, PATCH, DELETE) need the
    `X-CSRF-Token` header from `GET /auth/session` and pass the browser
    cross-origin check; request bodies must be `application/json`.
servers:
  - url: /api/v1
paths:
  /openapi.json:
    get:
      operationId: getOpenAPI
      x-meshsdr-access: anonymous
      summary: This OpenAPI document, as JSON
      tags: [meta]
      responses:
        "200":
          description: OpenAPI document
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: "#/components/responses/Problem"
...
  /auth/session:
    get:
      operationId: getSession
      x-meshsdr-access: anonymous
      summary: Current identity, roles, CSRF token and auth providers
      description: |
        Returns the caller's identity and the CSRF token to send in
        `X-CSRF-Token` on state-changing requests. Without a session, the
        token is bound to a pre-session cookie set by this response
        (login and other anonymous forms).
      tags: [auth]
      responses:
        "200":
          description: Current session (or anonymous)
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        default:
          $ref: "#/components/responses/Problem"
...
components:
...
  schemas:
...
    SessionInfo:
      type: object
      required: [authenticated, roles, csrf_token, providers]
      properties:
        authenticated:
          type: boolean
        user:
          $ref: "#/components/schemas/SessionUser"
        roles:
          type: array
          description: Roles held globally or on some device, lowest first. Empty when anonymous.
          items:
            type: string
            enum: [listener, operator, admin]
        csrf_token:
          type: string
          description: Value of the X-CSRF-Token header for state-changing requests.
        providers:
          type: array
          description: Available auth providers (`local` only in v1).
          items:
            type: string
    SessionUser:
      type: object
      required: [id, username, must_change_password]
      properties:
        id:
          type: string
          description: User id (UUID).
          example: 0192c3a4-5b6c-7d8e-9f01-23456789abcd
        username:
          type: string
        display_name:
          type: string
        must_change_password:
          type: boolean
...
    Problem:
      description: RFC 9457 problem details, served as application/problem+json.
      type: object
      required: [type, title, status, code]
      properties:
        type:
          type: string
          description: Always "about:blank"; the error is identified by `code`.
          example: about:blank
        title:
          type: string
          description: Short summary of the HTTP status.
          example: Not Found
        status:
          type: integer
          example: 404
        code:
          type: string
          description: Stable snake_case error code.
          example: not_found
        detail:
          type: string
          description: Human-readable explanation of this occurrence.
        errors:
          type: array
          description: Field errors of an invalid request.
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [path, code, message]
      properties:
        path:
          type: string
          example: username
        code:
          type: string
          example: required
        message:
          type: string
...
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"